// app/shared/schema.ts
export interface Database {
// ...
  public: {
    Tables: {
// ...
      teams: {
        Row: {
          description: string
          id: string
          name: string
        }
        Insert: {
          description: string
          id: string
          name: string
        }
        Update: {
          description?: string
          id?: string
          name?: string
        }
        Relationships: []
      }
      userteams: {
        Row: {
          teamid: string
          userid: string
        }
        Insert: {
          teamid: string
          userid: string
        }
        Update: {
          teamid?: string
          userid?: string
        }
        Relationships: [
          {
            foreignKeyName: "userteams_teamid_fkey"
            columns: ["teamid"]
            referencedRelation: "teams"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "userteams_userid_fkey"
            columns: ["userid"]
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      taskstatus: "ToDo" | "InProgress" | "InReview" | "Testing" | "Done"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}